# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mocker admin API",
    "version": "1.0.0",
    "description": "The admin plane of mocker: everything the web UI talks to.\n\nThis document is the CONTRACT, not documentation-after-the-fact. web/ generates its\nentire API client from it (orval), and internal/admin/openapi_contract_test.go fails\nthe Go build's test bar if a single (method, path) here disagrees with the routes\ninternal/admin actually registers. Adding a handler without adding it here — or the\nreverse — is a red test, not a silent drift.\n\nEvery route below is served on the admin host (MOCKER_ADMIN_HOST) only. The mock\nplane (per-workspace hosts) is a different surface entirely and is NOT described here."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as the UI: the SPA is served by this very binary, so the client never needs an absolute base URL."
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mocker_session",
        "description": "Set by POST /api/auth/login. HttpOnly, SameSite=Lax, and Secure unless MOCKER_DEV=1."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required on every POST/PUT/PATCH/DELETE except login. Value comes from the login/me response's csrfToken. Compared in constant time. Two further checks apply to the same requests and are NOT expressible here: Content-Type must parse to exactly application/json, and Origin (or Referer, when Origin is absent) must have hostname == MOCKER_ADMIN_HOST."
      }
    },
    "parameters": {
      "workspaceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Workspace id.",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
      "checkpointId": {
        "name": "cid",
        "in": "path",
        "required": true,
        "description": "Checkpoint id.",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "No session, or an expired one. The UI treats this on any non-login call as \"the session died\" and returns to the login screen.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "CSRF refusal: a missing/incorrect X-CSRF-Token, or an Origin/Referer whose hostname is not MOCKER_ADMIN_HOST.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "UnsupportedMediaType": {
        "description": "Content-Type did not parse to application/json.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The single error envelope every non-2xx answer on this plane uses (internal/httpx/respond.go).",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine code. The shared set is bad_request, unauthorized, forbidden, not_found, conflict, too_large, internal; session's GET/POST/DELETE add service_unavailable (no LiveStateSource wired in this deployment), and login adds rate_limited and unsupported_media_type. not_implemented_yet does not appear anywhere in this contract: it named DESIGN §12's scenario switch before scenarios existed (POST .../session {\"scenario\":...} answered 501), and stopped being true the moment scenarios shipped their own activate/deactivate routes (A17) — that key is a plain 400 here now. A3 adds edit_conflict on the five per-row compare-and-swap routes (PATCH /api/workspaces/{id}, PUT .../operations/{opKey}, POST .../auth-preset, PUT .../endpoints/{eid}, PUT .../scenarios/{sid}): the caller's editVersion expectation no longer matches the row's stored edit_version.",
                "examples": [
                  "bad_request",
                  "unauthorized",
                  "forbidden",
                  "not_found",
                  "conflict",
                  "too_large",
                  "internal",
                  "rate_limited",
                  "unsupported_media_type",
                  "service_unavailable",
                  "edit_conflict"
                ]
              },
              "message": {
                "type": "string",
                "description": "Human text. Shown to the operator as-is."
              },
              "details": {
                "description": "Per-field validation detail. Absent unless the route produces one."
              }
            }
          }
        }
      },
...
      "RollbackRequest": {
        "type": "object",
        "description": "DESIGN §14:871 draws restoreData; confirmSlug is P3d's own addition, required exactly when restoreData is true. restoreData:true restores the checkpoint's entity rows (data_snap) alongside the workspace layer — see POST .../rollback/{cid}'s own description for the refusals this can hit.",
        "properties": {
          "restoreData": {
            "type": "boolean",
            "description": "false or absent (default): the workspace layer (settings, overrides, custom endpoints) is restored; entity rows are read for this rollback's own pre-destructive checkpoint but never mutated. true (P3d): entity rows are restored too, from the target checkpoint's data_snap — requires confirmSlug and can be refused (409/413/500, see the route's own responses)."
          },
          "confirmSlug": {
            "type": "string",
            "description": "Required, and must equal the workspace's own slug (read inside the same transaction as the write), exactly when restoreData is true — the same check POST .../resource-decisions and POST .../reset-data already require for a destructive call. Ignored when restoreData is false or absent."
          }
        }
      },
      "RollbackResponseView": {
        "type": "object",
        "description": "The workspace layer was restored to the target checkpoint's state and a new revision was allocated (max+1; revision numbers are never reused).",
        "required": [
          "revision",
          "scenarioActive",
          "dataRestored"
        ],
        "properties": {
          "revision": {
            "type": "integer",
            "format": "int64",
            "description": "The workspace's revision AFTER this call."
          },
          "scenarioActive": {
            "type": "boolean",
            "description": "Whether a scenario is (still) active for this workspace after the rollback. Rollback is allowed while a scenario is active and never deactivates it — the screen should warn that any key the active scenario names still overrides what was just restored."
          },
          "dataRestored": {
            "type": "boolean",
            "description": "P3d's no-op signal for the entity-data half of the restore: true when the restore ran for at least one family the checkpoint carried, false when every carried family was skipped (its resources row no longer exists) or the checkpoint carried none, and always false when restoreData was false or absent. Not an echo of the request's own restoreData — a RESOLVED family whose stored and live relations are both empty still counts as restored."
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/workspaces/{id}/rollback/{cid}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/workspaceId"
        },
        {
          "$ref": "#/components/parameters/checkpointId"
        }
      ],
      "post": {
        "tags": [
          "checkpoints"
        ],
        "operationId": "rollbackWorkspace",
        "summary": "Restore the workspace layer to a checkpoint",
        "description": "Settings, operation overrides and custom endpoints are all restored wholesale (never merged) to the checkpoint's snapshot. The snapshot's confirmed-family configuration (resources, resource_decisions — P3b) is restored UPSERT-ONLY: a row the snapshot names is written back, but no row is ever deleted by this call, so a family confirmed after the checkpoint survives a rollback to it. restoreData:true (P3d) additionally restores the entity rows themselves from the checkpoint's data_snap: a family declined after the checkpoint comes back both configured AND populated, requires confirmSlug (checked inside the same transaction as everything else this call writes), and is refused rather than run partially when the target checkpoint carries no entity data, its stored document is unreadable, or this rollback's own pre-destructive capture of the CURRENT entity state degrades past the size cap — see the 409/413 responses below. Protected by a pre-destructive checkpoint of the CURRENT state (settings AND, since P3d, entity rows) written in the same transaction, and a new revision is allocated (max+1). Allowed while a scenario is active — see scenarioActive on the response.",
        "security": [
          {
            "sessionCookie": [],
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Restored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollbackResponseView"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body — including a body naming restoreData or confirmSlug with the wrong JSON type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "No such workspace, or no such checkpoint id in this workspace — including one that belongs to a DIFFERENT workspace (the two answer identically).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Five distinct meanings share this status, told apart by error.code: conflict — the workspace kept changing while it was being snapshotted, retry; confirm_slug_required — restoreData:true and confirmSlug was omitted or empty; confirm_slug_mismatch — restoreData:true and confirmSlug did not match the workspace's own slug, checked inside the same transaction as everything else this call writes; no_data_snapshot — restoreData:true against a checkpoint whose data_snap IS NULL (it predates P3d, or its own capture degraded). Every one of the four writes nothing. endpoint_ref_unresolved (P7a, D6) — the checkpoint's endpoint rows carry a `$ref` the workspace's CURRENT spec cannot resolve (a rollback restores layers, never the spec binding); details is a list of EndpointRefConflict rows, endpointId omitted, and nothing is written.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Three distinct meanings share this status, told apart by error.code: too_large — the CONFIGURATION half of this rollback's own pre-destructive snapshot (settings, overrides, custom endpoints) exceeds the checkpoint size cap; data_snapshot_too_large (P3d) — restoreData:true and the ENTITY-DATA half of that same pre-destructive snapshot degrades past its own cap, so unchecking restoreData is what would let the rollback through; too_large (httpx) — the request body itself exceeded the plane-wide limit before this route ever saw it. Every one writes nothing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "500": {
            "description": "Server-side failure. Most causes are generic (the message is deliberately vague; the detail is in the server log), but restoreData:true against a checkpoint whose data_snap decodes yet fails validation — a hand-built or otherwise corrupted document, never one this build itself writes — answers 500 rather than a 4xx, because nothing a client sent caused it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}